lmers::workflow

Module runner

Structs

AtomMapping

FormatOptions

RenameOptions

Enums

Runner

RunnerOutput

Statics

CACHED_READ_STACK

Cached static for the cached_read_stack function.

Functions

cached_read_stack

In a workflow, the base and existed layers will not be modified or deleted, so the result of read_stack function is in fact only dependent on the path parameter so create a cached function here is reasonable.

cached_read_stack_no_cache

Origin of the cached function cached_read_stack.

cached_read_stack_prime_cache

Primes the cached function cached_read_stack. In a workflow, the base and existed layers will not be modified or deleted, so the result of read_stack function is in fact only dependent on the path parameter so create a cached function here is reasonable.